move SSEHeaders to impl and make package protected
(cherry picked from commit d618fbbe6cac07fcc0e006f922cef57d56c107a6)
