+++ vertx-web/src/main/java/io/vertx/ext/web/handler/sse/EventSource.java
@@ -34,7 +34,6 @@ public interface EventSource {
   @Fluent
   EventSource connect(String path, Handler<AsyncResult<Void>> handler);
 
-  @Fluent
   default void close() {}
 
   @Fluent
